LIBASPACE-266. Updated ArchivesSpace config to match v2.7.0-1.0
Updated the "config/config.rb" and "config/plugins" file to match the
changes in aspace-env v2.7.0-1.0-rc1.

https://bitbucket.org/umd-lib/aspace-env/src/19aa1d02dbd780a6b3f86adc58790752102271b5

Note: The "config/config.rb" file is essentially a "merge" of the
v2.7.0-1.0-rc1 file with the elements that were in the aspace-docker
version. Elements that were commented out in the aspace-docker version
were commented out in this version.

https://issues.umd.edu/browse/LIBASPACE-266